requirements.txt
Flask
tensorflow
numpy
pillow
gdown
